requirements.txt
python-telegram-bot>=21.0,<22.0
openai>=1.60.0
python-dotenv>=1.0.1
